Updates UI to better inform the user what's going on
I added a pop-up window that appears during fitting to display the LL at each iteration.

I also updated hjcfit_20180130.m to accept a parameter for the pop-up window to allow
it to display the info at each iteration.

I also changed the Resolution and Concatenate buttons to sliders to try to make it clear
whether resolution is imposed (or not) and whether the dwells were concatenated (or not).
This also allows for a more defined workflow and an easy check at fitting time that these
two steps were done. It's also kinda nice to undo and redo these steps.