Added a script to read csv/txt
Hi So I use idls that are typically in csv format. So added an importer for that.
ps... Any relation to David O.?  If so please say "Hi" from one of his old students.
Also save people having to add each time, just a "prepare script" that does that if you run it from root.
Cool program.